+++ testkit/src/main/java/org/apache/calcite/test/CalciteAssert.java
@@ -1006,7 +1006,7 @@ public class CalciteAssert {
       // FOODMART, and this allows statistics queries to be executed.
       foodmart = addSchemaIfNotExists(rootSchema, SchemaSpec.JDBC_FOODMART);
       final Wrapper salesTable =
-          requireNonNull((Wrapper) foodmart.getTable("sales_fact_1997"));
+          requireNonNull((Wrapper) foodmart.tables().get("sales_fact_1997"));
       SchemaPlus fake =
           rootSchema.add(schema.schemaName, new AbstractSchema());
       fake.add("time_by_day", new AbstractTable() {
@@ -1079,7 +1079,7 @@ public class CalciteAssert {
 
   private static SchemaPlus addSchemaIfNotExists(SchemaPlus rootSchema,
         SchemaSpec schemaSpec) {
-    final SchemaPlus schema = rootSchema.getSubSchema(schemaSpec.schemaName);
+    final SchemaPlus schema = rootSchema.subSchemas().get(schemaSpec.schemaName);
     if (schema != null) {
       return schema;
     }
@@ -2272,11 +2272,11 @@ public class CalciteAssert {
       }
     };
 
-    @Override public Table getTable(String name) {
+    @Deprecated @Override public Table getTable(String name) {
       return table;
     }
 
-    @Override public Set<String> getTableNames() {
+    @Deprecated @Override public Set<String> getTableNames() {
       return ImmutableSet.of("myTable");
     }
 
@@ -2297,11 +2297,11 @@ public class CalciteAssert {
       return ImmutableSet.of();
     }
 
-    @Override public @Nullable Schema getSubSchema(String name) {
+    @Deprecated @Override public @Nullable Schema getSubSchema(String name) {
       return null;
     }
 
-    @Override public Set<String> getSubSchemaNames() {
+    @Deprecated @Override public Set<String> getSubSchemaNames() {
       return ImmutableSet.of();
     }